The treemap is alive with a caveat
The tree map now functions with a hiarichal fashion allowing all big boxes to represent the number of the movies.

NOTE: Right now the treemap is set to binary instead of squarify because it looks nicer, we may be able to play with ratio to fix a issue with sizing.